portableService: Add emptyFiles and emptyDirs parameters
In order to bind-mount files or folders, there needs to be an existing
(possibly-empty) corresponding file or folder inside the image.

When using, for example, the StateDirectory=myservice systemd service
option, systemd will create /var/lib/myservice on the host and then
attempt to bind-mount this into the filesystem. As it stands, this
currently fails because /var/lib/myservice doesn't exist.

This change enables the user to request additional files and folders,
which are then creates alongside the standard required files and folders.